changed uderId type to long

+++ src/main/java/com/banti/wallet/ums/service/UserService.java
@@ -4,7 +4,6 @@ import java.util.List;
 import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import com.banti.wallet.ums.model.User;
 import com.banti.wallet.ums.repository.UserRepository;
 
@@ -23,11 +22,11 @@ public class UserService {
         repo.save(user);
     }
      
-    public User get(Integer id) {
+    public User get(Long id) {
         return repo.findById(id).get();
     }
      
-    public void delete(Integer id) {
+    public void delete(Long id) {
         repo.deleteById(id);
     }
 }
